% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmPval.R
\name{kmPval}
\alias{kmPval}
\title{kmPval}
\usage{
kmPval(model)
}
\arguments{
\item{model}{is a survdiff() model object}
}
\value{
pvalue shown in survdiff() model as a character string
}
\description{
kmPval provides log-rank p-value provided in the survdiff() call.
}
\examples{
library(survival)
mod <- survdiff(Surv(runif(100, .1, 5),
                sample(c(0, 1), 100, replace = T, prob = c(.2, .8))
                ) ~ sample(c('no', 'yes'), 100, replace = T))
kmPval(mod)
}

