% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerPvals.R
\name{lmerPvals}
\alias{lmerPvals}
\title{lmerPvals}
\usage{
lmerPvals(model)
}
\arguments{
\item{model}{A lmer() model object}
}
\value{
returns the model coefficients with the approximated p-vales appended
}
\description{
lmerPvals provides p-values for lmer() models using the normal distribution as an approximation
}

