% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_boot.R
\name{logistic_boot}
\alias{logistic_boot}
\title{logistic_boot}
\usage{
logistic_boot(data, formula, R = 10000, seed = 12345)
}
\arguments{
\item{data}{data.frame containing variables to be modeled}

\item{formula}{is a formula object specifying the model}

\item{R}{is the number of random samples}

\item{seed}{seed to be set for replicability}
}
\value{
A data.frame with the original OR and the bootstrapped 95% CIs
}
\description{
logistic_boot provides bootstrapped  with 95% confidence intervals odds ratios
}

