% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marascuiloPHA.R
\name{marascuiloPHA}
\alias{marascuiloPHA}
\title{marascuiloPHA}
\usage{
marascuiloPHA(data, group, var, pval = 0.05, dec = 3)
}
\arguments{
\item{data}{data.frame containing all data}

\item{group}{Variable name of the grouping variable}

\item{var}{Variance name for the variable of interest}

\item{pval}{Significance level. Default of 0.05}

\item{dec}{Number of decimals to round descriptive statistics to.}
}
\value{
A list of matrices.
}
\description{
marascuiloPHA performs the Marascuilo PHA procedure, which is a conservative chi-square pairwise comparison tests. Specifically, given a level of a grouping variable of interest, this procedure will test whether or not the proportion of patients that are that level differ pairwise across all levels of another variable.
}
\examples{
example_df <- data.frame(x1 = sample(letters[1:3], 30, replace = T),
                         x2 = sample(letters[10:14], 30, replace = T))
example_df
with(example_df, table(x1, x2))
marascuiloPHA(data = example_df, group = "x1", var = "x2")
}

