% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseTable.R
\name{pairwiseTable}
\alias{pairwiseTable}
\title{pairwiseTable}
\usage{
pairwiseTable(levels, printOutput = FALSE,
  documentName = "PairwiseTable.doc")
}
\arguments{
\item{levels}{A Vector of levels to use for the pairwise table}

\item{printOutput}{If TRUE, pairwise table will be printed to a word document}

\item{documentName}{If printOutput==T, the name of the output word document}
}
\description{
pairwiseTable provides a table explicating all pairwise comparisons made, along with their pairwise number found in the multGroup function.
}
\examples{
levels<-c("White", "Black", "Other")
pairwiseTable(levels)
}

