% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthocChiSq.R
\name{posthocChiSq}
\alias{posthocChiSq}
\title{posthocChiSq}
\usage{
posthocChiSq(chiObject, control = c("none", "fdr", "BH", "BY", "bonferroni",
  "holm", "hochberg", "hommel"), digits = 4)
}
\arguments{
\item{digits}{}
}
\value{
data.frame containing raw and adjusted p-values.
}
\description{
posthocChiSq computes standard pairwise chi-squared tests and provides raw and adjusted p-values.
}
\examples{
example_df <- data.frame(x1 = sample(c('m', 'f'), 50, replace = T),
                         x2 = sample(letters[1:4], 50, replace = T))
test <- with(example_df, chisq.test(x2, x1))
posthocChiSq(chiObject = test)
test2 <- with(example_df, chisq.test(x1, x2))
posthocChiSq(chiObject = test2, control="holm")
}

