% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictClustMembership.R
\name{predictClustMembership}
\alias{predictClustMembership}
\title{predictClustMembership}
\usage{
predictClustMembership(clusterObject, newData)
}
\arguments{
\item{clusterObject}{R object from kmeans()}

\item{newData}{data.frame containing the same variables as were used for computing the clusters in the `clusterObject`}
}
\value{
Vector of cluster membership assignment
}
\description{
predictClustMembership will predict/assign cluster membership to new data based on a kmeans() cluster object
}
\examples{
df1 <- data.frame(x= rnorm(100), x2 = rnorm(100), x3 = rnorm(100))
mod <- kmeans(df1, 2)
df2 <- data.frame(x= rnorm(100), x2 = rnorm(100), x3 = rnorm(100))
predictClustMembership(mod, df2)
}

