% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrroom.R
\name{purrroom}
\alias{purrroom}
\title{purrroom}
\usage{
purrroom(mapData, broomCall = function(.) {     tidy(.) })
}
\arguments{
\item{mapData}{a purrr call that ends with a mapped model call}

\item{broomCall}{a broom call (e.g. glance(), tidy()) call wrapped in a function with a dot argument}
}
\value{
Tidy data.frame containing all split model summaries.
}
\description{
purrroom easily integrates purrr::map modeling into broom. Providing the user with a single tidy dataframe for all models.
}
\examples{
mapDat <-
mtcars \%>\%
  split(.$cyl) \%>\%
  map(~ lm(mpg ~ wt, data = .))
## As a function call:
broomCall <- function(.) {tidy(., conf.int=T)}
purrroom(mapDat, broomCall)
## Using piping:
mtcars \%>\%
  split(.$cyl) \%>\%
  map(~ lm(mpg ~ wt, data = .)) \%>\%
  purrroom()
}

